+++ DesignToolbox/DesignToolbox/Pages/Components/Checkbox/CheckboxItem/TwoStates/CheckboxItemElement.swift
@@ -11,6 +11,7 @@
 // Software description: A SwiftUI components library with code examples for Orange Unified Design System
 //
 
+import OUDS
 import SwiftUI
 
 /// Related to `OUDSCheckboxItem` (i.e. with 2 available values).
@@ -26,6 +27,7 @@ struct CheckboxItemElement: DesignToolboxElement {
             name: name,
             image: nil,
             description: "app_components_checkbox_controlItem_description_text",
+            version: OUDSVersions.componentCheckboxVersion,
             illustration: AnyView(CheckboxItemPage())))
     }
 }